+++ spec/dummy/app/mailers/test_mailer.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class TestMailer < ApplicationMailer
-  def send_simple_email
+  def send_simple_email(*)
     mail(
       to: 'user@gmail.com',
       subject: 'Test email',